more g

+++ dataset/CoSiM-commented-by-Gemini/raw/298d196a-b97c-47c9-98bb-b716996c3c26/gpu_hashtable.cu
@@ -1,5 +1,12 @@
-
-
+/**
+ * @file gpu_hashtable.cu
+ * @brief CUDA implementation of a GPU-resident hash table.
+ *
+ * This file defines a hash table data structure that is stored and manipulated
+ * entirely on the GPU. It uses an open addressing scheme with linear probing for
+ * collision resolution. Operations like insertion and retrieval are performed in
+ * parallel batches by CUDA kernels.
+ */
 
 #include 
 #include 
@@ -12,33 +19,54 @@
 
 #include "gpu_hashtable.hpp"
 
-
+// Defines the maximum load factor before the hash table is resized.
 #define LOAD_FACTOR                     0.8f
+// Constants for operation success/failure.
 #define FAIL                            false
 #define SUCCESS                         true
 
 
-
+/**
+ * @brief Host-side constructor for the GpuHashTable.
+ * @param size The initial number of buckets to allocate for the hash table.
+ *
+ * Allocates memory on the GPU for the hash table buckets and initializes
+ * all entries to zero, marking them as empty.
+ */
  GpuHashTable::GpuHashTable(int size) {
     limitSize = size;
     currentSize = 0;
+    // Allocate device memory for the hash buckets.
     cudaMalloc(&hashTableBuckets, limitSize * sizeof(HashTableEntry));
 
     if (hashTableBuckets == 0) {
         cerr << "[HOST] Couldn't allocate memory for GpuHashTable!\n";
     }
 
+    // Initialize all bucket keys to KEY_INVALID (0).
     cudaMemset(hashTableBuckets, 0, limitSize * sizeof(HashTableEntry));
 
 }
 
-
+/**
+ * @brief Host-side destructor for the GpuHashTable.
+ *
+ * Frees the GPU memory allocated for the hash table buckets.
+ */
 GpuHashTable::~GpuHashTable() {
     cudaFree(hashTableBuckets);
 }
 
 
-
+/**
+ * @brief A device function to compute the hash of an integer key.
+ * @param data The input key to hash.
+ * @param limit The size of the hash table, used for the modulo operation.
+ * @return The computed hash value, an index into the hash table.
+ *
+ * This function uses a series of bitwise operations and multiplications
+ * (a variant of MurmurHash) to create a well-distributed hash value.
+ */
 __device__ int getHash(int data, int limit) {
     
     data ^= data >> 16;
@@ -51,11 +79,25 @@ __device__ int getHash(int data, int limit) {
 }
 
 
-
+/**
+ * @brief CUDA kernel for inserting a batch of key-value pairs into the hash table.
+ * @param keys Pointer to the device array of keys.
+ * @param values Pointer to the device array of values.
+ * @param numKeys The number of key-value pairs to insert.
+ * @param hashTableBuckets Pointer to the hash table buckets on the device.
+ * @param limitSize The total number of buckets in the hash table.
+ *
+ * Each thread in the grid processes one key-value pair. It calculates the hash,
+ * then uses linear probing and an atomic Compare-and-Swap (CAS) operation to
+ * find an empty bucket or update an existing key. The atomic CAS is critical
+ * for ensuring thread safety during concurrent insertions.
+ */
 __global__ void kernelInsertEntry(int *keys, int *values, int numKeys, HashTableEntry *hashTableBuckets, int limitSize) {
 
+    // Determine the unique ID for the current thread.
     int threadId = blockIdx.x * blockDim.x + threadIdx.x;
   
+    // Boundary check: ensure the thread is not processing out of bounds.
     if (threadId >= numKeys)
         return;
     
@@ -63,23 +105,39 @@ __global__ void kernelInsertEntry(int *keys, int *values, int numKeys, HashTable
     int currentValue = values[threadId];
     int hash = getHash(currentKey, limitSize);
     
+    // Begin linear probing loop.
     while(true) {
-
-
+        // Atomically attempt to write the currentKey into an empty bucket (marked by KEY_INVALID).
+        // `atomicCAS` returns the old value that was in the bucket.
         int inplaceKey = atomicCAS(&hashTableBuckets[hash].HashTableEntryKey, KEY_INVALID, currentKey);
+        
+        // If the old value was KEY_INVALID, this thread successfully claimed the bucket.
+        // If the old value was the same as currentKey, we are updating an existing entry.
         if (inplaceKey == currentKey || inplaceKey == KEY_INVALID) {
             hashTableBuckets[hash].HashTableEntryValue = currentValue;
             return;
         }
+        // If the bucket was occupied by another key, move to the next bucket.
         hash = (hash + 1) % limitSize;
     }
 }
 
 
-
+/**
+ * @brief Host-side function to insert a batch of key-value pairs.
+ * @param keys Pointer to the host array of keys.
+ * @param values Pointer to the host array of values.
+ * @param numKeys The number of pairs to insert.
+ * @return `true` on success, `false` on failure.
+ *
+ * This function orchestrates the batch insertion process, including checking the
+ * load factor, resizing if necessary, transferring data to the GPU, launching
+ * the kernel, and cleaning up.
+ */
  bool GpuHashTable::insertBatch(int *keys, int* values, int numKeys) {
-    int futureLoadFactor = (float) (currentSize + numKeys) / limitSize;
+    float futureLoadFactor = (float) (currentSize + numKeys) / limitSize;
    
+    // If adding the new keys exceeds the load factor, resize the table first.
     if (futureLoadFactor >= LOAD_FACTOR) {
         reshape(limitSize + numKeys);
     }
@@ -87,6 +145,7 @@ __global__ void kernelInsertEntry(int *keys, int *values, int numKeys, HashTable
     int *deviceKeys;
     int *deviceValues;
 
+    // Allocate memory on the device for the keys and values.
     cudaMalloc(&deviceKeys, numKeys * sizeof(int));
     cudaMalloc(&deviceValues, numKeys * sizeof(int));
 
@@ -95,14 +154,17 @@ __global__ void kernelInsertEntry(int *keys, int *values, int numKeys, HashTable
         return FAIL;
     }
 
+    // Copy the key and value data from host to device.
     cudaMemcpy(deviceKeys, keys, numKeys * sizeof(int), cudaMemcpyHostToDevice);
     cudaMemcpy(deviceValues, values, numKeys * sizeof(int), cudaMemcpyHostToDevice);
 
+    // Configure the kernel launch parameters (grid and block size).
     int threadBlockSize = 1024;
     int gridSize = numKeys/ threadBlockSize;
     if (numKeys % threadBlockSize)
         gridSize++;
 
+    // Launch the insertion kernel.
     kernelInsertEntry>>(
             deviceKeys,
             deviceValues,
@@ -111,19 +173,32 @@ __global__ void kernelInsertEntry(int *keys, int *values, int numKeys, HashTable
             limitSize
             );
 
+    // Wait for the kernel to complete and check for errors.
     cudaError_t cudaerr = cudaDeviceSynchronize();
     if (cudaerr != cudaSuccess)
         printf("kernel launch failed with error \"%s\".\n", cudaGetErrorString(cudaerr));
 
     currentSize += numKeys;
 
+    // Free the temporary device memory.
     cudaFree(deviceKeys);
     cudaFree(deviceValues);
    
 	return SUCCESS;
 }
 
-
+/**
+ * @brief CUDA kernel for retrieving a batch of values corresponding to a batch of keys.
+ * @param keys Pointer to the device array of keys to look up.
+ * @param values Pointer to the device array where the results will be stored.
+ * @param numKeys The number of keys to look up.
+ * @param limitSize The size of the hash table.
+ * @param hashTableBuckets Pointer to the hash table buckets on the device.
+ *
+ * Each thread processes one key. It uses the same linear probing logic as the
+ * insert kernel to find the key. If found, the corresponding value is written
+ * to the output array; otherwise, a sentinel value (0) is written.
+ */
 __global__ void kernelGetEntry( int *keys, int *values, int numKeys, int limitSize, HashTableEntry *hashTableBuckets) {
 
     int threadId = blockIdx.x * blockDim.x + threadIdx.x;
@@ -133,30 +208,39 @@ __global__ void kernelGetEntry( int *keys, int *values, int numKeys, int limitSi
     int currentKey = keys[threadId];
     int hash = getHash(currentKey, limitSize);
     
+    // Linear probing loop to find the key.
     while(true) {
-
-
+        // If the key is found, write the value to the output array and terminate.
         if (hashTableBuckets[hash].HashTableEntryKey == currentKey) {
             values[threadId] =  hashTableBuckets[hash].HashTableEntryValue;
             return;
         }
 
+        // If an empty bucket is encountered, the key does not exist in the table.
         if (hashTableBuckets[hash].HashTableEntryKey == KEY_INVALID) {
-            values[threadId] = 0;
+            values[threadId] = 0; // Sentinel for 'not found'.
             return;
         }
 
+        // Move to the next bucket.
         hash = (hash + 1) % limitSize;
     }
 }
 
 
-
+/**
+ * @brief Host-side function to retrieve a batch of values.
+ * @param keys Pointer to the host array of keys to look up.
+ * @param numKeys The number of keys to retrieve.
+ * @return A pointer to a host array containing the corresponding values.
+ *         The caller is responsible for freeing this memory. Returns NULL on failure.
+ */
  int* GpuHashTable::getBatch(int* keys, int numKeys) {
     int *deviceKeys;
     int *values;
     int *deviceValues;
 
+    // Allocate memory on the device for the input keys and output values.
     cudaMalloc(&deviceKeys, numKeys * sizeof(int));
     cudaMalloc(&deviceValues, numKeys * sizeof(int));
 
@@ -165,8 +249,10 @@ __global__ void kernelGetEntry( int *keys, int *values, int numKeys, int limitSi
         return NULL;
     }
 
+    // Copy keys from host to device.
     cudaMemcpy(deviceKeys, keys, numKeys * sizeof(int), cudaMemcpyHostToDevice);
 
+    // Configure and launch the get kernel.
     int threadBlockSize = 1024;
     int gridSize = numKeys/ threadBlockSize;
     if (numKeys % threadBlockSize)
@@ -180,30 +266,42 @@ __global__ void kernelGetEntry( int *keys, int *values, int numKeys, int limitSi
             hashTableBuckets
             );
 
+    // Wait for the kernel to finish.
     cudaError_t cudaerr = cudaDeviceSynchronize();
     if (cudaerr != cudaSuccess)
         printf("kernel launch failed with error \"%s\".\n", cudaGetErrorString(cudaerr));
-        
+    
+    // Allocate host memory for the results.
     values = (int *) calloc(numKeys, sizeof(int));
     if (values == NULL) {
         cerr << "[HOST] Couldn't allocate memory for thre return values array!\n";
         return NULL;
     }
 
+    // Copy the results from device to host.
     cudaMemcpy(values, deviceValues, numKeys * sizeof(int), cudaMemcpyDeviceToHost);
 
+    // Clean up device memory.
     cudaFree(deviceValues);
     cudaFree(deviceKeys);
 
     return values;
 }
 
-
+/**
+ * @brief CUDA kernel to copy elements from an old hash table to a new, larger one.
+ * @param hashTableBucketsOrig Pointer to the original (old) hash table.
+ * @param limitSizeOrig The size of the original hash table.
+ * @param hashTableBuckets Pointer to the new, resized hash table.
+ * @param limitSize The size of the new hash table.
+ *
+ * This kernel is used for rehashing. Each thread takes an element from the old
+ * table and re-inserts it into the new table using the same atomic linear
+ * probing logic as the main insertion kernel.
+ */
 __global__ void kernelCopyTable(
         HashTableEntry *hashTableBucketsOrig,
         int limitSizeOrig,
-
-
         HashTableEntry *hashTableBuckets,
         int limitSize) {
 
@@ -211,16 +309,17 @@ __global__ void kernelCopyTable(
     if (threadId >= limitSizeOrig)
         return;
 
+    // Skip empty buckets from the old table.
     if (hashTableBucketsOrig[threadId].HashTableEntryKey == KEY_INVALID)
         return;
 
     int currentKey = hashTableBucketsOrig[threadId].HashTableEntryKey;
     int currentValue = hashTableBucketsOrig[threadId].HashTableEntryValue;
+    // Re-hash the key for the new table size.
     int hash = getHash(currentKey, limitSize);
     
+    // Use atomic linear probing to insert into the new table.
     while (true) {
-
-
         int inplaceKey = atomicCAS(&hashTableBuckets[hash].HashTableEntryKey, KEY_INVALID, currentKey);
         if (inplaceKey == currentKey || inplaceKey == KEY_INVALID) {
             hashTableBuckets[hash].HashTableEntryValue = currentValue;
@@ -230,12 +329,16 @@ __global__ void kernelCopyTable(
     }
 }
 
-
+/**
+ * @brief Resizes the hash table to a new capacity and rehashes all existing elements.
+ * @param numBucketsReshape The new number of buckets for the resized table.
+ */
 void GpuHashTable::reshape(int numBucketsReshape) {
-    cout << "pennis"<< endl;
+    cout << "pennis"<< endl; // Note: Unprofessional debug output.
     HashTableEntry *hashTableBucketsReshaped;
     int newLimitSize = numBucketsReshape;
 
+    // Allocate a new, larger table on the device.
     cudaMalloc(&hashTableBucketsReshaped, newLimitSize * sizeof(HashTableEntry));
 
     if (hashTableBucketsReshaped == 0)
@@ -243,6 +346,7 @@ void GpuHashTable::reshape(int numBucketsReshape) {
 
     cudaMemset(hashTableBucketsReshaped, 0, newLimitSize * sizeof(HashTableEntry));
 
+    // Configure and launch the kernel to copy/rehash elements.
     int threadBlockSize = 1024;
     int gridSize = limitSize / threadBlockSize;
     if (limitSize % threadBlockSize)
@@ -255,17 +359,23 @@ void GpuHashTable::reshape(int numBucketsReshape) {
             newLimitSize
             );
 
+    // Wait for the copy to complete.
     cudaError_t cudaerr = cudaDeviceSynchronize();
     if (cudaerr != cudaSuccess)
         printf("kernel launch failed with error \"%s\".\n", cudaGetErrorString(cudaerr));
     
+    // Free the old table memory.
     cudaFree(hashTableBuckets);
     
+    // Point the class members to the new table and update the size.
     hashTableBuckets = hashTableBucketsReshaped;
     limitSize = newLimitSize;
 }
 
-
+/**
+ * @brief Calculates and returns the current load factor of the hash table.
+ * @return The load factor as a float.
+ */
 float GpuHashTable::loadFactor() {
     if (currentSize == 0)
         return 0.f;
@@ -274,7 +384,7 @@ float GpuHashTable::loadFactor() {
 }
 
 
-
+// Convenience macros for using the hash table in the test file.
 #define HASH_INIT GpuHashTable GpuHashTable(1);
 #define HASH_RESERVE(size) GpuHashTable.reshape(size);
